Add support for revoking OTP with mail
The API added support to revoke OTP using a mail token. This commit
updates the front accordingly.
API changes have also been applied here, this adds the `/revoke` route
which works the same as the one to reset or unlock account.

The URL format for the `/unlock`, `/reinitialize` and `/revoke` pages have
been changed. Since we now have three way to authenticate, the `totp`
boolean in the URL have been replaced by a `type` string which indicates
"password", "totp" or "token". This changes applies to the 3 routes
since the `askModal.vue` was modified. Note: this will break existing
token URLs sent by mail.